feat: add auth, user, role, permission, color, size, product API

+++ src/main/resources/db/migration/V2__create_roles_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE roles
 (
-    id          BIGINT AUTO_INCREMENT PRIMARY KEY,
+    id          INT AUTO_INCREMENT PRIMARY KEY,
     name        VARCHAR(100) NOT NULL UNIQUE,
     description VARCHAR(255),
